+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/manager/DeveloperManager.java
@@ -136,9 +136,9 @@ public class DeveloperManager extends SecuredManager {
     public DeveloperTree getHierarchyById(Long id) throws EntityRetrievalException {
         List<CertificationBodyDTO> acbs = acbManager.getAll();
         Developer developer = getById(id);
-        List<Product> products = productManager.getByDeveloper(developer.getDeveloperId());
-        List<ProductVersionDTO> versions = versionManager.getByDeveloper(developer.getDeveloperId());
-        List<CertifiedProductDetailsDTO> listings = certifiedProductDao.findListingsByDeveloperId(developer.getDeveloperId());
+        List<Product> products = productManager.getByDeveloper(developer.getId());
+        List<ProductVersionDTO> versions = versionManager.getByDeveloper(developer.getId());
+        List<CertifiedProductDetailsDTO> listings = certifiedProductDao.findListingsByDeveloperId(developer.getId());
 
         DeveloperTree developerTree = new DeveloperTree(developer);
         products.stream().forEach(product -> {
@@ -148,7 +148,7 @@ public class DeveloperManager extends SecuredManager {
         developerTree.getProducts().stream().forEach(product -> {
             List<ProductVersionDTO> productVersions =
                     versions.stream()
-                    .filter(version -> version.getProductId().equals(product.getProductId()))
+                    .filter(version -> version.getProductId().equals(product.getId()))
                     .collect(Collectors.toList());
             productVersions.stream().forEach(version -> {
                 product.getVersions().add(new VersionTree(version));
@@ -158,7 +158,7 @@ public class DeveloperManager extends SecuredManager {
         developerTree.getProducts().stream().forEach(product -> {
             product.getVersions().stream().forEach(version -> {
                 List<SimpleListing> listingsForVersion = listings.stream()
-                        .filter(listing -> listing.getVersion().getId().equals(version.getVersionId()))
+                        .filter(listing -> listing.getVersion().getId().equals(version.getId()))
                         .map(listing -> convertToSimpleListing(listing, acbs))
                         .collect(Collectors.toList());
                     version.getListings().addAll(listingsForVersion);
@@ -206,7 +206,7 @@ public class DeveloperManager extends SecuredManager {
     }, allEntries = true)
     public Developer update(Developer updatedDev, boolean doUpdateValidations)
             throws EntityRetrievalException, JsonProcessingException, EntityCreationException, ValidationException {
-        Developer beforeDev = getById(updatedDev.getDeveloperId());
+        Developer beforeDev = getById(updatedDev.getId());
 
         if (updatedDev.equals(beforeDev)) {
             LOGGER.info("Developer did not change - not saving");
@@ -233,8 +233,8 @@ public class DeveloperManager extends SecuredManager {
         }
         developerDao.update(updatedDev);
         updateStatusHistory(beforeDev, updatedDev);
-        Developer after = getById(updatedDev.getDeveloperId());
-        activityManager.addActivity(ActivityConcept.DEVELOPER, after.getDeveloperId(),
+        Developer after = getById(updatedDev.getId());
+        activityManager.addActivity(ActivityConcept.DEVELOPER, after.getId(),
                 "Developer " + updatedDev.getName() + " was updated.", beforeDev, after);
         return after;
     }
@@ -266,13 +266,13 @@ public class DeveloperManager extends SecuredManager {
 
             if (hasChanged) {
                 DeveloperStatusEvent dseToUpdate = toUpdate.getUpdated();
-                dseToUpdate.setDeveloperId(beforeDev.getDeveloperId());
+                dseToUpdate.setDeveloperId(beforeDev.getId());
                 developerDao.updateDeveloperStatusEvent(dseToUpdate);
             }
         }
 
         for (DeveloperStatusEvent toAdd : statusEventsToAdd) {
-            toAdd.setDeveloperId(beforeDev.getDeveloperId());
+            toAdd.setDeveloperId(beforeDev.getId());
             developerDao.createDeveloperStatusEvent(toAdd);
         }
 
@@ -299,7 +299,7 @@ public class DeveloperManager extends SecuredManager {
         }
 
         Long developerId = developerDao.create(developer);
-        developer.setDeveloperId(developerId);
+        developer.setId(developerId);
 
         Developer createdDevloperDto = developerDao.getById(developerId);
         activityManager.addActivity(ActivityConcept.DEVELOPER, developerId,