+++ chpl/chpl-service/src/test/java/gov/healthit/chpl/permissions/domain/changerequest/UpdateActionPermissionsTest.java
@@ -48,7 +48,7 @@ public class UpdateActionPermissionsTest extends ActionPermissionsBaseTest {
 
         Mockito.when(changeRequestDAO.get(ArgumentMatchers.anyLong()))
                 .thenReturn(ChangeRequest.builder()
-                        .developer(Developer.builder().developerId(2L).build())
+                        .developer(Developer.builder().id(2L).build())
                         .build());
     }
 
@@ -62,20 +62,20 @@ public class UpdateActionPermissionsTest extends ActionPermissionsBaseTest {
 
         Mockito.when(changeRequestDAO.get(ArgumentMatchers.anyLong()))
                 .thenReturn(ChangeRequest.builder()
-                        .developer(Developer.builder().developerId(2L).build())
+                        .developer(Developer.builder().id(2L).build())
                         .build());
 
         assertTrue(permissions.hasAccess(ChangeRequest.builder()
-                .developer(Developer.builder().developerId(2L).build())
+                .developer(Developer.builder().id(2L).build())
                 .build()));
 
         Mockito.when(changeRequestDAO.get(ArgumentMatchers.anyLong()))
                 .thenReturn(ChangeRequest.builder()
-                        .developer(Developer.builder().developerId(3l).build())
+                        .developer(Developer.builder().id(3L).build())
                         .build());
 
         assertFalse(permissions.hasAccess(ChangeRequest.builder()
-                .developer(Developer.builder().developerId(3l).build())
+                .developer(Developer.builder().id(3L).build())
                 .build()));
 
     }